Handle non-Debian based hosts
This role is a depdendency in some roles that are run on non-Debian based
hosts, so this role must handle those cases. I have added an empty var file
and set lookup() to default to that file to handle any non-specified case as
a no-op. I have re-added Amazon Linux 2 and Fedora 32/33 as supported hosts
as well.