Add an option to update the package cache
Prior to #41 the role would update the package cache after adding a
backports repository. This restores that functionality as an option
with a default of `false` to preserve the current role behavior. This
will allow a consuming role or playbook to update the cache without
adding an explicitly task to do so.