Adjust package cache updating logic
Change the default value of backports_update_cache to `true`. This
reflects what a role consumer would likely expect and stays true to the
default value of `update_cache` for the ansible.builtin.apt_repository
module. The `update_cache` value for the pre-DEB822 task is now
supplied by the backports_update_cache variable. The logic for
distributions that use DEB822 SourceLists is modified to update the
cache both when adding a repository results in a change in addition to
the value of the backports_update_cache variable. This change also
allows us to remove the idempotence skipping tag from the package cache
update task. Lastly the update_cache molecule scenario is changed to
provide a value of `false` for backports_update_cache.

Co-authored-by: Shane Frasier <jeremy.frasier@gwe.cisa.dhs.gov>